requirements.txt
kafka-python==2.0.2
pandas==2.0.2
Flask==2.3.2
apache-flink==1.17.1